requirements.txt
asgiref==3.6.0
Django==4.2.1
django-bootstrap5==23.1
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
soupsieve==2.4.1
sqlparse==0.4.3
whitenoise==6.4.0
environs[django]==9.5.0
django-tinymce==3.6.1